backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
stripe==7.10.0
SQLAlchemy==1.4.51
python-dotenv==1.0.0
gunicorn==21.2.0
